langgraph-basics/pyproject.toml
[project]
name = "test-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.0",
    "langchain-google-genai>=4.1.2",
    "langchain-groq>=1.1.1",
    "langgraph>=1.0.5",
    "langgraph-checkpoint-mongodb>=0.3.0",
    "openai>=2.9.0",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
